+++ src/Indicators/t-z/ZigZag/ZigZag.Catalog.cs
@@ -11,7 +11,7 @@ public static partial class ZigZag
             .WithId("ZIGZAG")
             .WithCategory(Category.PriceTransform)
             .AddEnumParameter<EndType>("endType", "End Type", description: "Type of price to use for the calculation", isRequired: false, defaultValue: EndType.Close)
-            .AddParameter<double>("percentChange", "Percent Change", description: "Minimum percent change required for a new direction", isRequired: false, defaultValue: 5.0, minimum: 1.0, maximum: 200.0)
+            .AddParameter<decimal>("percentChange", "Percent Change", description: "Minimum percent change required for a new direction", isRequired: false, defaultValue: 5.0m, minimum: 1.0, maximum: 200.0)
             .AddResult("ZigZag", "Zig Zag", ResultType.Default, isReusable: true)
             .AddResult("PointType", "Point Type", ResultType.Default)
             .AddResult("RetraceHigh", "Retrace High", ResultType.Default)
